trading/requirements.txt
# Trading module requirements
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
httpx>=0.24.0
python-dotenv>=1.0.0

